options consolidation
